+++ packages/core-node/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/core-node
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- fix package installation
+
 ## 1.0.0-beta.7
 
 ### Patch Changes