+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/test-utils
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- @pipelab/plugin-core@1.0.0-beta.8
+
 ## 1.0.0-beta.7
 
 ### Patch Changes