+++ plugins/plugin-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pipelab/plugin-core
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @pipelab/core-node@1.0.0-beta.8
+
 ## 1.0.0-beta.7
 
 ### Patch Changes